+++ src/components/input/input.js
@@ -1,4 +1,4 @@
-import './index.css'
+import './input.css'
 
 const Input = (props) => {
     return <input type='text' {...props} />
